ai-core/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
aiohttp==3.9.1
python-dotenv==1.0.0
PyPDF2==3.0.1
httpx==0.25.2
asyncio==3.4.3
logging==0.4.9.6
